requirements.txt
streamlit
streamlit-cropper
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic
chromadb
pypdf2
python-docx
pandas
matplotlib
seaborn
scikit-learn
umap-learn
openai
python-dotenv
pillow
openpyxl
pytesseract
opencv-python
unstructured[all-docs]
